requirement.txt
streamlit==1.24.0
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.0
pydantic==2.0.3